Try loading bundles unspecified in the master file
This makes possible adding bundles just by populating their locale-specific
directies - without modifying the master file, which may not be easy after
the product has been deployed and the language packs cannot modify it.

Bundles are attempted to load on the first access and the success or failure
are remembered by true or false in the master object not to repeat the loading
attemot later again.